extract logger and spsc queue to momo common
