+++ momo_common/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(momo_common INTERFACE)
+
+target_include_directories(momo_common INTERFACE
+        ${CMAKE_CURRENT_SOURCE_DIR}/include
+)
