+++ momo_common/include/Envelope.hpp
