+++ momo_feed/src/config/ConfigLoader.cpp
@@ -3,9 +3,6 @@
 #include <yaml-cpp/yaml.h>
 #include <filesystem>
 
-#include "common/JsonLogger.hpp"
-#include "common/SPSCQueue.hpp"
-
 namespace fs = std::filesystem;
 
 template <typename T>
