+++ momo_feed/src/futu_quote/FutuQuoteSession.cpp
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include <iostream>
 
-#include "common/JsonLogger.hpp"
+#include "../../../momo_common/include/JsonLogger.hpp"
 
 FutuQuoteSession::FutuQuoteSession(SessionCallbacks cbs) : cbs_(cbs) {
     qot_ = new Futu::FTAPI_Qot();
